ngfw-13258 added alert-tests test suite
